Use short ContentNames also for SI and PI
TS 102 818 specifies SI and PI XML-names.
However TS 102 371 mandates that while converting that XML to a binary format, the information from the original XML-name should be passed to the MOT Object attributes like ScopeId, ScopeStart and ScopeEnd. ContentName is not the original filename. It is a reference to the MOT directory and can be a short name.